fix: edge node interface changes ignored(CSD-105)
So the issue is that zedcloud changes the interfaces list ordering. Zedcloud is transforms interfaces from a list to a map and stores it as a map in the database. So, there is no way to restore ordering without changing how we store data in the database.  To suppress this, we have DiffSuppresFunc defined: https://github.com/zededa/terraform-provider-zedcloud/blob/main/v2/schemas/node.go#L671. This function allows the `terraform plan` to not complain about the interfaces ordering, but doesn't show changes in the existing elements of the interfaces list.

The best practice in Terraform is to use `Type: schema.TypeSet` instead of `Type: schema.TypeList` for fields where the ordering of elements is not essential or where we can’t have consistent ordering.
The problem with the schema.TypeSet is that when we make changes in one of the elements of the list or add a new element, the terraform plan will show that it needs to update the whole interfaces list, but it is a proper behavior by terraform as it treats interfaces as one object.
```
# zedcloud_edgenode.ivan_test_tf_provider will be updated in-place
  ~ resource "zedcloud_edgenode" "ivan_test_tf_provider" {
        id                        = "6c3d0fb5-983a-4138-a42f-f2224dee9aa9"
        name                      = "ivan_test_tf_provider"
        tags                      = {
            "tag-key-1" = "tag-value-1"
            "tag-key-2" = "tag-value-2"
        }
        # (21 unchanged attributes hidden)

      - interfaces {
          - cost       = 0 -> null
          - intf_usage = "ADAPTER_USAGE_UNSPECIFIED" -> null
          - intfname   = "eth0" -> null
          - tags       = {} -> null
        }
      - interfaces {
          - cost       = 255 -> null
          - intf_usage = "ADAPTER_USAGE_MANAGEMENT" -> null
          - intfname   = "defaultIPv4" -> null
          - ipaddr     = "127.0.0.1" -> null
          - macaddr    = "00:00:00:00:00:00" -> null
          - tags       = {
              - "system_interface_1_key" = "system_interface_1_value"
              - "system_interface_2_key" = "system_interface_2_value"
            } -> null
        }
      + interfaces {
          + cost       = 0
          + intf_usage = "ADAPTER_USAGE_MANAGEMENT"
          + intfname   = "COM3"
        }
      + interfaces {
          + cost       = 255
          + intf_usage = "ADAPTER_USAGE_MANAGEMENT"
          + intfname   = "defaultIPv4"
          + ipaddr     = "127.0.0.1"
          + macaddr    = "00:00:00:00:00:00"
          + tags       = {
              + "system_interface_1_key" = "system_interface_1_value"
              + "system_interface_2_key" = "system_interface_2_value"
            }
        }
      + interfaces {
          + intf_usage = "ADAPTER_USAGE_MANAGEMENT"
          + intfname   = "eth0"
        }

        # (3 unchanged blocks hidden)
    }

```